Move from package `io.luizotavio.minecraft` into `io.github.luizotavio.placeholder4j` and resolve issues related to some placeholders
